# api/openapi.yaml
openapi: 3.0.3
info:
  title: StudyWiki-Agent API
  version: 1.0.0
  description: |
    StudyWiki-Agent 服务契约 —— 基于 LangChain + LangGraph 的可增长本地 Wiki 知识库 API。

    统一约定:
    - 成功响应通常为 `{"status":"success","data":...}`（部分端点直接返回该字典，部分经 `ApiResponse` 包装并额外携带 `message`/`error_code`/`timestamp`）。
    - 错误响应统一为 `{"status":"error","error_code":"SW-...","message":"...","timestamp":"..."}`。
    - 除 `/health` 返回 `{"status":"ok","cards_count":N}` 外，业务接口顶层均含 `status` 字段。

    鉴权:
    - 当 `.env` 配置 `STUDYWIKI_AUTH_TOKEN` 时启用，业务 API 需携带 `Authorization: Bearer <token>`。
    - public 路径（无需鉴权）: `/`、`/health`、`/docs`、`/openapi.json`、`/api/openapi.yaml`、`/api/bootstrap/status`、`/api/bootstrap/test`、`/api/bootstrap/configure`，以及 `/static/*`。
    - 未授权返回 401，错误码 `SW-AUTH-001`。

    > 本文档为手写契约，与 `api/openapi.yaml` 一致；实际运行时的自动导出见 `/openapi.json`。
servers:
  - url: http://127.0.0.1:8000
    description: 本地开发服务器
...
paths:
...
  /api/logs:
    get:
      tags: [health]
      summary: 获取最近运行日志
      parameters:
        - name: level
          in: query
          required: false
          schema: { type: string }
          description: 可选日志级别过滤
        - name: n
          in: query
          required: false
          schema: { type: integer, default: 100 }
      responses:
        "200":
          description: 最近日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, enum: [success] }
                  data:
                    type: array
                    items: { type: string }
  /api/openapi.yaml:
    get:
      tags: [health]
      summary: 返回手写 OpenAPI 契约
      security: []
      responses:
        "200":
          description: 本契约文档（YAML）
          content:
            application/yaml:
              schema: { type: string }
        "404":
          $ref: "#/components/responses/Error"

  # ── Bootstrap ──────────────────────────────────────────
...
  /api/bootstrap/test:
    post:
      tags: [bootstrap]
      summary: 验证 API Key（不落盘）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapConfigRequest" }
      responses:
        "200":
          description: "验证结果（status 为 success 或 error）"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/BootstrapTestResult" }
        "400":
          $ref: "#/components/responses/Error"
  /api/bootstrap/configure:
    post:
      tags: [bootstrap]
      summary: 验证成功后写入 .env 并解除灰屏
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BootstrapConfigRequest" }
      responses:
        "200":
          description: "配置结果（status 为 success 或 error）"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/BootstrapTestResult" }
        "400":
          $ref: "#/components/responses/Error"

  # ── Cards ──────────────────────────────────────────────
  /api/cards:
    get:
      tags: [cards]
      summary: 列出卡片（分页 + 分类过滤）
      parameters:
        - name: category
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          required: false
          schema: { type: integer, default: 1, minimum: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, default: 50, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: 卡片列表
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CardListResponse" }
    post:
      tags: [cards]
      summary: 创建卡片
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardCreate" }
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CardResponse" }
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/cards/search:
    get:
      tags: [cards]
      summary: 语义搜索卡片
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 1 }
      responses:
        "200":
          description: 搜索结果
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          cards:
                            type: array
                            items: { $ref: "#/components/schemas/CardResponse" }
                          total: { type: integer }
  /api/cards/generate:
    post:
      tags: [cards]
      summary: LLM 生成知识卡片内容
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardCreate" }
      responses:
        "201":
          description: 生成并创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CardResponse" }
        "500":
          $ref: "#/components/responses/Error"
...
  /api/cards/{card_id}:
    get:
      tags: [cards]
      summary: 获取单张卡片
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: 卡片详情
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CardResponse" }
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [cards]
      summary: 更新卡片（部分字段）
      parameters:
        - $ref: "#/components/parameters/CardId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CardUpdate" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data: { $ref: "#/components/schemas/CardResponse" }
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [cards]
      summary: 删除卡片
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: 删除成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
        "404":
          $ref: "#/components/responses/Error"

  # ── Categories ─────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 当 STUDYWIKI_AUTH_TOKEN 配置时启用；public 路径无需携带。
  parameters:
    CardId:
      name: card_id
      in: path
      required: true
      schema: { type: string }
      description: 卡片 ID
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: 请求校验失败（FastAPI 默认结构）
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
  schemas:
    ApiResponse:
      type: object
      description: 统一 API 响应格式
      required: [status]
      properties:
        status:
          type: string
          enum: [success, error]
        data:
          type: object
          nullable: true
        message:
          type: string
          default: ""
        error_code:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      description: 统一错误响应体
      required: [status, error_code, message, timestamp]
      properties:
        status:
          type: string
          enum: [error]
        error_code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorCode:
      type: string
      description: |
        统一错误码，格式 `SW-<DOMAIN>-<CODE>`。
        数值后缀通常对应该错误建议的 HTTP 状态码。
      enum:
        - SW-AUTH-001
        - SW-CARD-400
        - SW-CARD-404
        - SW-CARD-500
        - SW-UPLOAD-400
        - SW-UPLOAD-413
        - SW-UPLOAD-500
        - SW-QUIZ-400
        - SW-QUIZ-404
        - SW-QUIZ-500
        - SW-TASK-404
        - SW-SETTINGS-400
        - SW-KB-400
        - SW-KB-404
        - SW-KB-500
        - SW-BOOTSTRAP-400
        - SW-BOOTSTRAP-401
        - SW-BOOTSTRAP-429
        - SW-BOOTSTRAP-UPSTREAM
        - SW-BOOTSTRAP-TIMEOUT
        - SW-BOOTSTRAP-NETWORK
        - SW-AGENT-400
        - SW-AGENT-429
        - SW-LLM-503
        - SW-BACKUP-404
        - SW-BACKUP-500
        - SW-PAIR-400
        - SW-PAIR-401
        - SW-PAIR-404
        - SW-DB-507
        - SW-GENERIC-400
        - SW-GENERIC-401
        - SW-GENERIC-403
        - SW-GENERIC-404
        - SW-GENERIC-405
        - SW-GENERIC-409
        - SW-GENERIC-413
        - SW-GENERIC-422
        - SW-GENERIC-500
        - SW-GENERIC-502
        - SW-GENERIC-503
        - SW-GENERIC-504
...
    CardCreate:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        aliases:
          type: array
          items: { type: string }
          default: []
        content: { type: string, default: "" }
        examples:
          type: array
          items: { type: string }
          default: []
        questions:
          type: array
          items: { type: string }
          default: []
        category: { type: string, default: 未分类 }
        source_file: { type: string, default: "" }
        source_page: { type: integer, default: 0 }
        related_cards:
          type: array
          items: { type: string }
          default: []
    CardUpdate:
      type: object
      description: 更新卡片请求（所有字段可选）
      properties:
        title: { type: string, nullable: true }
        aliases:
          type: array
          items: { type: string }
          nullable: true
        content: { type: string, nullable: true }
        examples:
          type: array
          items: { type: string }
          nullable: true
        questions:
          type: array
          items: { type: string }
          nullable: true
        category: { type: string, nullable: true }
        source_file: { type: string, nullable: true }
        source_page: { type: integer, nullable: true }
        related_cards:
          type: array
          items: { type: string }
          nullable: true
    CardResponse:
      type: object
      description: 卡片响应
      required: [id, title, category, created_at, updated_at]
      properties:
        id: { type: string }
        title: { type: string }
        aliases:
          type: array
          items: { type: string }
        content: { type: string }
        examples:
          type: array
          items: { type: string }
        questions:
          type: array
          items: { type: string }
        category: { type: string }
        source_file: { type: string }
        source_page: { type: integer }
        related_cards:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CardListResponse:
      type: object
      properties:
        cards:
          type: array
          items: { $ref: "#/components/schemas/CardResponse" }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    BootstrapConfigRequest:
      type: object
      required: [api_key]
      properties:
        provider:
          type: string
          enum: [deepseek, openai]
          default: deepseek
        api_key:
          type: string
          minLength: 8
          maxLength: 512
        base_url:
          type: string
          maxLength: 512
          default: ""
        model:
          type: string
          maxLength: 128
          default: ""
...
    BootstrapTestResult:
      type: object
      properties:
        ok: { type: boolean }
        provider: { type: string }
        key_tail: { type: string }
...